// src/saju/enums/twelve-life-stages.enum.ts
/**
 * 12운성(十二運星) - 일간의 생명력 상태를 나타내는 12단계
 */
export enum TwelveLifeStages {
  JANG = '장생',    // 長生 - 태어남, 시작
  MOK = '목욕',     // 沐浴 - 씻김, 정화
  GWAN = '관대',    // 冠帶 - 성장, 교육
  IM = '임관',      // 臨官 - 건록, 성숙
  JE = '제왕',      // 帝旺 - 왕성함, 절정
  SUE = '쇠',       // 衰 - 쇠퇴 시작
  BYEONG = '병',    // 病 - 병듦, 약화
  SA = '사',        // 死 - 죽음, 정지
  MYO = '묘',       // 墓 - 무덤, 저장
  JEOL = '절',      // 絶 - 절멸, 소멸
  TAE = '태',       // 胎 - 태아, 잉태
  YANG = '양'       // 養 - 양육, 준비
}